fixed bugs + textChangedLister

+++ app/src/main/java/com/example/nlapp/model/CurrencyRate.kt
@@ -1,7 +1,7 @@
 package com.example.nlapp.model
 
 data class CurrencyRate(
-    val amount: Int?,
+    val amount: Long?,
     val from: String?,
     val to: String?,
     val value: Double?
